Do not select future ephemeris
When selecting ephemeris to use for an observation, avoid selecting an ephemeris
whose Toe is after the time of the observation.

This can be useful for constellations such as Galileo, where the ephemeris are
usually broadcast by the satellites starting several minutes after their Toe and
do not give a good fit if used some time (an hour, for example) before their Toe.